Bug: 347933 Description: merging Sun and Mozilla.org liblber Fix Description: 1) remove some casts for arguments to ber_scanf/ber_printf/ber_put*/ber_get* that are not needed anymore e.g. don't need to cast bv_val from long to int anymore (better still would be to use 'O') 2) found a couple of places that were using long * instead of int *.  I'm surprised these didn't cause problems on 64 bit platforms. 3) Use 0x%p instead of 0x%x or 0x%lx to print an address (e.g. LDAP *).  I guess 0x%lx would work on on 32bit and 64bit, but %p is portable and proper. 4) changed ldapmsg.lm_msgtype from int to ber_tag_t 5) changed lr_res_msgtype from int to ber_tag_t

+++ mozilla/directory/c-sdk/ldap/libraries/libldap/control.c
@@ -108,7 +108,7 @@ nsldapi_put_controls( LDAP *ld, LDAPControl **ctrls, int closeseq,
 
 		if ( c->ldctl_value.bv_val != NULL ) {
 			if ( ber_printf( ber, "o", c->ldctl_value.bv_val,
-			    (int)c->ldctl_value.bv_len /* XXX lossy cast */ )
+			    c->ldctl_value.bv_len )
 			    == -1 ) {
 				goto error_exit;
 			}
